server/requirements.txt
openenv-core>=0.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
# Security: pin minimum safe versions for known CVEs
starlette>=0.49.1
cryptography>=46.0.5
# Perceptual hashing for reverse image search (server/ml/phash.py)
# Pillow pulls in libjpeg/zlib — see Dockerfile apt-get install
pillow>=10.0.0
